fix failing tests
