+++ hedera-node/src/main/java/com/hedera/services/store/contracts/precompile/utils/PrecompilePricingUtils.java
@@ -145,7 +145,7 @@ public class PrecompilePricingUtils {
                         .setSignedTransactionBytes(signedTxn.toByteString())
                         .build();
 
-        TxnAccessor accessor;
+        SignedTxnAccessor accessor;
         try{
             accessor = accessorFactory.constructSpecializedAccessor(txn.toByteArray());
         } catch (InvalidProtocolBufferException e) {
