+++ hedera-node/src/test/java/com/hedera/services/utils/accessors/custom/TokenWipeAccessorTest.java
@@ -13,7 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hedera.services.utils.accessors;
+package com.hedera.services.utils.accessors.custom;
+
+import com.google.protobuf.InvalidProtocolBufferException;
+import com.hedera.services.context.properties.GlobalDynamicProperties;
+import com.hederahashgraph.api.proto.java.TokenWipeAccountTransactionBody;
+import com.hederahashgraph.api.proto.java.TransactionBody;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.util.List;
 
 import static com.hedera.services.utils.accessors.SignedTxnAccessorTest.buildTransactionFrom;
 import static com.hedera.test.utils.IdUtils.asAccount;
@@ -24,17 +35,6 @@ import static com.hederahashgraph.api.proto.java.SubType.TOKEN_FUNGIBLE_COMMON;
 import static com.hederahashgraph.api.proto.java.SubType.TOKEN_NON_FUNGIBLE_UNIQUE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.BDDMockito.given;
-
-import com.google.protobuf.InvalidProtocolBufferException;
-import com.hedera.services.context.properties.GlobalDynamicProperties;
-import com.hedera.services.utils.accessors.custom.TokenWipeAccessor;
-import com.hederahashgraph.api.proto.java.TokenWipeAccountTransactionBody;
-import com.hederahashgraph.api.proto.java.TransactionBody;
-import java.util.List;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class TokenWipeAccessorTest {
